# swaggerdoc/src/app/api/model/animalPart.ts
/**
 * JWTToken_Auth_API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type AnimalPart = 1 | 2 | 3 | 4;

export const AnimalPart = {
    NUMBER_1: 1 as AnimalPart,
    NUMBER_2: 2 as AnimalPart,
    NUMBER_3: 3 as AnimalPart,
    NUMBER_4: 4 as AnimalPart
};