# swaggerdoc/src/app/api/model/intentionDto.ts
/**
 * JWTToken_Auth_API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AnimalPart } from './animalPart';
import { IntentionStatus } from './intentionStatus';


export interface IntentionDto { 
    id?: number;
    code?: string | null;
    name: string;
    userId?: string | null;
    cityId?: number | null;
    message?: string | null;
    intentionStatus?: IntentionStatus;
    animalPart?: AnimalPart;
}
export namespace IntentionDto {
}