# swaggerdoc/src/app/api/model/intentionStatus.ts
/**
 * JWTToken_Auth_API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type IntentionStatus = 2 | 4 | 6 | 7;

export const IntentionStatus = {
    NUMBER_2: 2 as IntentionStatus,
    NUMBER_4: 4 as IntentionStatus,
    NUMBER_6: 6 as IntentionStatus,
    NUMBER_7: 7 as IntentionStatus
};